// src/app/modules/user-management/constants/User-edit.interface.ts
export enum UserEditFields {
  FULL_NAME = "fullName",
  PASS_WORD = "passWord",
  PHONE = "phone",
  GENDER = "gender",
  BIRTH_DAY = "birthDay",
  CITY = "city",
  WARD = "ward",
  DISTRICT = "district",
  ADDRESS = "address",
}
export interface UsersEditDTO {
  [UserEditFields.FULL_NAME]?: string;
  [UserEditFields.PASS_WORD]?: string;
  [UserEditFields.PHONE]?: string;
  [UserEditFields.GENDER]?: string;
  [UserEditFields.BIRTH_DAY]?: string;
  [UserEditFields.CITY]?: string;
  [UserEditFields.WARD]?: string;
  [UserEditFields.DISTRICT]?: string;
  [UserEditFields.ADDRESS]?: string;
}